pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-psycopg2"
version = "0.1.0"
description = "MCP server exposing psycopg2 PostgreSQL database adapter functionality"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Darío Clavijo", email = "clavijodario@gmail.com"}
]
dependencies = [
    "psycopg2-binary>=2.9.11,<3.0.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "hatch",
    "pip-api",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-asyncio",
    "hypothesis",
]
lint = [
    "ruff",
    "mypy",
]
all = ["mcp-psycopg2[dev,test,lint]"]

[project.urls]
Homepage = "https://github.com/daedalus/mcp-psycopg2"
Repository = "https://github.com/daedalus/mcp-psycopg2"
Issues = "https://github.com/daedalus/mcp-psycopg2/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_psycopg2"]

[tool.hatch.build.targets.sdist]
include = ["src/mcp_psycopg2"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "ANN", "TCH", "N", "C4", "ARG"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ANN", "ARG"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
no_implicit_optional = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src --cov-fail-under=75"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude = ["tests/*", "*/__init__.py"]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
